move P3 to to punch hole moved punch hole to mod
