+++ src-tauri/src/network/mod.rs
@@ -1,6 +1,5 @@
 use std::collections::HashMap;
 use std::io::{Error, ErrorKind};
-use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr};
 use std::sync::Mutex;
 
 use smoke::messages::RoomId;
@@ -9,18 +8,17 @@ use smoke::{PubKey, User};
 use tauri::EventHandler;
 
 use tokio::io::BufReader;
-use tokio::net::UdpSocket;
 use tokio::select;
 use tokio::sync::{mpsc, oneshot};
 
 use tokio_kcp::{KcpConfig, KcpStream};
 
-use log::{error, trace};
+use log::{trace};
 
-use self::p3::ping_pong_peng;
+use self::holepunch::punch_hole;
 
 pub mod ctrl_chnl;
-mod p3;
+mod holepunch;
 
 pub const ENV: Config = Config {
   public_key: dotenv!("PUBLIC_KEY"),
@@ -86,11 +84,6 @@ pub async fn hole_punch(
 
   /* Holepunch using rhizome */
   let socket = punch_hole(ENV.server_address, &room_id.0).await?;
-
-  if let Err(err) = ping_pong_peng(&socket).await {
-    error!("P3 failure: {}", err);
-    return Err(tauri::Error::Io(Error::new(ErrorKind::Other, "P3 failure")));
-  }
 
   let mut stream = if other_key.as_ref() < ENV.public_key.as_bytes() {
     trace!("client mode");
@@ -170,76 +163,4 @@ pub fn chat_exists(state: tauri::State<'_, Networking>, id: RoomId) -> bool {
     Ok(chats) => chats.contains_key(&id),
     Err(_) => false,
   }
-}
-
-/** Create a new socket and holepunch it! */
-async fn punch_hole<A>(server_addr: A, ident: &[u8]) -> Result<UdpSocket, Error>
-where
-  A: tokio::net::ToSocketAddrs,
-{
-  trace!("initiating ip exchange");
-  // Create, bind, and connect the socket:
-  let socket = UdpSocket::bind("0.0.0.0:0").await?;
-  socket.connect(server_addr).await?;
-
-  // Send the server our identity (Used to match us with a peer)
-  socket.send(ident).await?;
-
-  // Wait for the server to send us a peer:
-  let mut b = [0u8; 512];
-  let size = socket.recv(&mut b).await?;
-
-  if size == 0 {
-    return Err(Error::new(
-      ErrorKind::UnexpectedEof,
-      "rhizome didn't transmit peer address",
-    ));
-  }
-
-  // Try parse the recieved peer address.
-  let addr = parse_addr(&b, size)?;
-
-  trace!("connecting to peer: {}", &addr);
-
-  // Swap the connection from the server to the peer.
-  socket.connect(addr).await?;
-
-  Ok(socket)
-}
-
-/** Parse a collection of bytes to a valid IP address. */
-fn parse_addr(b: &[u8; 512], size: usize) -> Result<SocketAddr, Error> {
-  // Parse the bytes into a valid socket address:
-  let ip = match b[0] {
-    4 => {
-      if size < 7 {
-        error!("parsing sock addr: '{:x?}' failed", &b[..size]);
-        return Err(Error::new(ErrorKind::InvalidData, "peer address malformed"));
-      }
-      IpAddr::V4(Ipv4Addr::new(b[1], b[2], b[3], b[4]))
-    }
-    6 => {
-      if size < 19 {
-        error!("parsing sock addr: '{:x?}' failed", &b[..size]);
-        return Err(Error::new(ErrorKind::InvalidData, "peer address malformed"));
-      }
-      IpAddr::V6(Ipv6Addr::from([
-        b[1], b[2], b[3], b[4], b[5], b[6], b[7], b[8], b[9], b[10], b[11], b[12], b[13], b[14],
-        b[15], b[16],
-      ]))
-    }
-    _ => {
-      error!("parsing peer addr: '{:x?}' failed", &b[..size]);
-      return Err(Error::new(ErrorKind::InvalidData, "peer address malformed"));
-    }
-  };
-
-  // Parse the remaining bytes to a valid port number:
-  let port = match b[0] {
-    4 => ((b[5] as u16) << 8) | b[6] as u16,
-    6 => ((b[17] as u16) << 8) | b[18] as u16,
-    _ => unreachable!(),
-  };
-
-  Ok(SocketAddr::new(ip, port))
 }
